fixed task list screen; optimized-refactored core task/user fetching; TODO: finish task list screen

+++ lib/core/providers/task_provider.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:tmanager/core/models/task_model.dart';
-import 'package:tmanager/core/service/task_service.dart';
+import 'package:tmanager/core/services/task_service.dart';
 import 'package:uuid/uuid.dart';
 
 String? getCurrentUserId() {
@@ -18,22 +18,33 @@ class TaskProvider with ChangeNotifier {
   List<TaskModel> _filteredTasks = [];
   List<TaskModel> get filteredTasks => _filteredTasks;
 
-  Future<void> loadTasks(String userId) async {
+  Future<void> loadTasks() async {
+    final userId = getCurrentUserId();
+    if (userId == null) return;
+
+    _isLoading = true;
+    notifyListeners();
+
     _tasks = await _taskService.loadTasks(userId);
     _filteredTasks = _tasks;
+
+    _isLoading = false;
     notifyListeners();
   }
 
-  Future<void> saveTask(String userId, TaskModel task) async {
+  Future<void> saveTask(TaskModel task) async {
+    final userId = getCurrentUserId();
+    if (userId == null) return;
+
     if (task.id == null) {
       await addTask(task.title, task.description);
     } else {
-      await updateTask(userId, task);
+      await _taskService.updateTask(userId, task);
     }
   }
 
   Future addTask(String title, String description) async {
-    final userId = getCurrentUserId(); // Get the current user ID
+    final userId = getCurrentUserId();
     if (userId != null) {
       final newTask = TaskModel(
         id: const Uuid().v4(),
@@ -46,7 +57,10 @@ class TaskProvider with ChangeNotifier {
     }
   }
 
-  Future<void> updateTask(String userId, TaskModel task) async {
+  Future<void> updateTask(TaskModel task) async {
+    final userId = getCurrentUserId();
+    if (userId == null) return;
+
     final taskIndex = _tasks.indexWhere((e) => e.id == task.id);
     if (taskIndex >= 0) {
       _tasks[taskIndex] = task;
@@ -55,7 +69,10 @@ class TaskProvider with ChangeNotifier {
     }
   }
 
-  Future<void> deleteTask(String userId, TaskModel task) async {
+  Future<void> deleteTask(TaskModel task) async {
+    final userId = getCurrentUserId();
+    if (userId == null) return;
+
     _tasks.remove(task);
     _filteredTasks = _tasks;
     await _taskService.deleteTask(userId, task.id!);
